Renaming CoreContextHandler to CoreAppContext and associated files/dirs/modules.
