+++ jetty-core/jetty-bom/pom.xml
@@ -52,7 +52,7 @@
       </dependency>
       <dependency>
         <groupId>org.eclipse.jetty</groupId>
-        <artifactId>jetty-core-webapp</artifactId>
+        <artifactId>jetty-coreapp</artifactId>
         <version>12.1.0-SNAPSHOT</version>
       </dependency>
       <dependency>
