--- jetty-core/jetty-core-webapp/src/main/config/etc/jetty-static-deploy.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0"?><!DOCTYPE Configure PUBLIC "-//Jetty//Configure//EN" "https://jetty.org/configure_10_0.dtd">
-
-<!-- =============================================================== -->
-<!-- Configure the "static" environment deployment defaults          -->
-<!-- =============================================================== -->
-<Configure>
-  <Ref refid="deploymentScanner">
-    <Call name="configureEnvironment">
-      <Arg>static</Arg>
-      <Set name="defaultContextHandlerClassName">
-        <Property name="jetty.deploy.defaultContextHandlerClass" default="org.eclipse.jetty.core.webapp.StaticContextHandler" />
-      </Set>
-    </Call>
-  </Ref>
-</Configure>
