--- jetty-core/jetty-core-webapp/src/test/java/org/eclipse/jetty/core/webapp/AbstractCleanEnvironmentTest.java
@@ -1,32 +1,0 @@
-//
-// ========================================================================
-// Copyright (c) 1995 Mort Bay Consulting Pty Ltd and others.
-//
-// This program and the accompanying materials are made available under the
-// terms of the Eclipse Public License v. 2.0 which is available at
-// https://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
-// which is available at https://www.apache.org/licenses/LICENSE-2.0.
-//
-// SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
-// ========================================================================
-//
-
-package org.eclipse.jetty.core.webapp;
-
-import org.eclipse.jetty.util.component.Environment;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.parallel.Isolated;
-
-/**
- * Ensure that each test is using a clean (empty) Environment,
- * by running isolated.
- */
-@Isolated
-public abstract class AbstractCleanEnvironmentTest
-{
-    @AfterEach
-    public void removeAllEnvironments()
-    {
-        Environment.removeAll();
-    }
-}
