+++ jetty-core/jetty-coreapp/src/test/resources/jetty-logging.properties
@@ -1,0 +1,5 @@
+# Jetty Logging using jetty-slf4j-impl
+#org.eclipse.jetty.deploy.DeploymentTempDirTest.LEVEL=DEBUG
+#org.eclipse.jetty.deploy.LEVEL=DEBUG
+#org.eclipse.jetty.util.Scanner=DEBUG
+#org.eclipse.jetty.io.LEVEL=WARN
