--- jetty-core/jetty-tests/jetty-core-test-webapps/jetty-core-bad-init-webapp/pom.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>org.eclipse.jetty.tests</groupId>
-    <artifactId>jetty-core-test-webapps</artifactId>
-    <version>12.1.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>jetty-core-bad-init-webapp</artifactId>
-  <packaging>jar</packaging>
-  <name>Core WebApp :: Tests :: Bad Init WebApp</name>
-  <description>Jetty Core WebApp :: Test WebApps :: Bad Init WebApp</description>
-
-  <properties>
-    <bundle-symbolic-name>${project.groupId}.tests.coreapp.badinit</bundle-symbolic-name>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.eclipse.jetty</groupId>
-      <artifactId>jetty-server</artifactId>
-      <scope>provided</scope>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-assembly-plugin</artifactId>
-        <executions>
-          <execution>
-            <id>create-core-webapp</id>
-            <goals>
-              <goal>single</goal>
-            </goals>
-            <phase>package</phase>
-            <configuration>
-              <descriptors>
-                <descriptor>src/main/assembly/core-webapp.xml</descriptor>
-              </descriptors>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-    </plugins>
-  </build>
-</project>
