--- jetty-core/jetty-tests/jetty-core-test-webapps/jetty-core-demo-webapp/pom.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>org.eclipse.jetty.tests</groupId>
-    <artifactId>jetty-core-test-webapps</artifactId>
-    <version>12.1.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>jetty-core-demo-webapp</artifactId>
-  <packaging>jar</packaging>
-  <name>Core WebApp :: Tests :: Demo WebApp</name>
-  <description>Jetty Core WebApp :: Test WebApps :: Demo WebApp</description>
-
-  <properties>
-    <bundle-symbolic-name>${project.groupId}.tests.coreapp.demo</bundle-symbolic-name>
-    <maven.deploy.skip>true</maven.deploy.skip>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.eclipse.jetty</groupId>
-      <artifactId>jetty-server</artifactId>
-      <scope>provided</scope>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-assembly-plugin</artifactId>
-        <executions>
-          <execution>
-            <id>create-core-webapp</id>
-            <goals>
-              <goal>single</goal>
-            </goals>
-            <phase>package</phase>
-            <configuration>
-              <descriptors>
-                <descriptor>src/main/assembly/core-webapp.xml</descriptor>
-              </descriptors>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-    </plugins>
-  </build>
-
-</project>
