+++ jetty-core/jetty-tests/jetty-test-coreapps/jetty-test-coreapp-bad-init/src/main/assembly/core-webapp.xml
