+++ jetty-core/jetty-tests/jetty-test-coreapps/jetty-test-coreapp-bad-init/src/main/java/org/example/ExampleBadSetServerHandler.java
