+++ jetty-core/jetty-tests/jetty-test-coreapps/jetty-test-coreapp-demo/src/main/assembly/core-webapp.xml
