+++ jetty-core/jetty-tests/jetty-test-coreapps/pom.xml
@@ -1,27 +1,25 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
   <modelVersion>4.0.0</modelVersion>
   <parent>
-    <groupId>org.eclipse.jetty</groupId>
-    <artifactId>jetty-core</artifactId>
+    <groupId>org.eclipse.jetty.tests</groupId>
+    <artifactId>jetty-tests</artifactId>
     <version>12.1.0-SNAPSHOT</version>
   </parent>
-  <groupId>org.eclipse.jetty.tests</groupId>
-  <artifactId>jetty-tests</artifactId>
-  <version>12.1.0-SNAPSHOT</version>
+
+  <artifactId>jetty-test-coreapps</artifactId>
   <packaging>pom</packaging>
-  <name>Core :: Tests</name>
+
+  <name>Core WebApp :: Test WebApps</name>
+  <description>Jetty Core WebApp :: Test WebApps</description>
 
   <modules>
-    <module>jetty-test-client-transports</module>
-    <module>jetty-test-coreapps</module>
-    <module>jetty-test-jmx</module>
-    <module>jetty-test-rfc</module>
+    <module>jetty-test-coreapp-bad-init</module>
+    <module>jetty-test-coreapp-demo</module>
+    <module>jetty-test-coreapp-http2-client</module>
   </modules>
 
   <properties>
     <maven.deploy.skip>true</maven.deploy.skip>
   </properties>
-
 </project>
