+++ jetty-core/pom.xml
@@ -17,7 +17,7 @@
     <module>jetty-bom</module>
     <module>jetty-client</module>
     <module>jetty-compression</module>
-    <module>jetty-core-webapp</module>
+    <module>jetty-coreapp</module>
     <module>jetty-deploy</module>
     <module>jetty-ee</module>
     <module>jetty-fcgi</module>
