+++ tests/test-distribution/test-distribution-common/src/test/java/org/eclipse/jetty/tests/distribution/DeployerTest.java
@@ -197,7 +197,7 @@ public class DeployerTest extends AbstractJettyHomeTest
             String testXmlStr = """
                 <?xml version="1.0"?>
                 <!DOCTYPE Configure PUBLIC "-//Jetty//Configure//EN" "https://jetty.org/configure.dtd">
-                <Configure class="org.eclipse.jetty.core.webapp.CoreContextHandler">
+                <Configure class="org.eclipse.jetty.core.app.CoreAppContext">
                   <Set name="contextPath">/demo</Set>
                   <Set name="displayName">Demo of Core Deploy WebApp</Set>
                 </Configure>
