+++ tests/test-distribution/test-distribution-common/src/test/java/org/eclipse/jetty/tests/distribution/DistributionTests.java
@@ -935,7 +935,7 @@ public class DistributionTests extends AbstractJettyHomeTest
             Files.writeString(phpXML, """
                 <?xml version="1.0"?>
                 <!DOCTYPE Configure PUBLIC "-//Jetty//Configure//EN" "https://jetty.org/configure_10_0.dtd">
-                <Configure class="org.eclipse.jetty.core.webapp.CoreContextHandler">
+                <Configure class="org.eclipse.jetty.core.app.CoreAppContext">
                   <Set name="contextPath">/php</Set>
                   <Set name="baseResourceAsPath">
                     <Call class="java.nio.file.Path" name="of">
